+++ Shared/Sources/Document/Artboard/Artboard.swift
@@ -2,17 +2,6 @@ import CoreGraphics
 
 import Foundation
 
-/// Data transfer object that represents file structure
-public struct FrameInfo: Codable {
-    public var id: UUID
-    public var imageScale: CGFloat
-    public var frame: CGRect
-    
-    public static var `default`: Self {
-        Self(id: UUID(), imageScale: 1, frame: .zero)
-    }
-}
-
 public class Artboard {
 //    let figmaURL: String
     public var frames: [Frame] = []
