+++ Shared/Sources/Document/Artboard/DescriptionExtensions/A11yDescription+Adjustable.swift
