+++ Shared/Sources/Document/Artboard/DescriptionExtensions/A11yDescription+Colors.swift
