+++ Shared/Sources/Document/Artboard/Protocols/AccessibilityView/AccessibilityView+ExtractElements.swift
