+++ Shared/Sources/Document/Artboard/Protocols/Parts/AdjustableOptions+Localization.swift
