+++ Shared/Sources/Document/Artboard/Protocols/Parts/AdjustableOptions.swift
