--- Shared/Sources/Document/Models/OptionalDescriptionSubject.swift
@@ -1,3 +1,0 @@
-import Combine
-
-public typealias OptionalDescriptionSubject = CurrentValueSubject<(any AccessibilityView)?, Never>
